requirements.txt
requests
beautifulsoup4
supabase
psycopg2-binary
dotenv